Rename xml_checker to schema_checker
