+++ qc_openscenario/checks/__init__.py
@@ -1,2 +1,3 @@
-from . import xml_checker as xml_checker
+from . import schema_checker as schema_checker
+from . import basic_checker as basic_checker
 from . import models as models
