+++ qc_openscenario/checks/schema_checker/schema_constants.py
@@ -1,0 +1,1 @@
+CHECKER_ID = "schema_xosc"
