--- qc_openscenario/checks/xml_checker/__init__.py
@@ -1,3 +1,0 @@
-from . import xml_constants as xml_constants
-from . import xml_checker as xml_checker
-from . import schema_is_valid as schema_is_valid
