--- qc_openscenario/checks/xml_checker/xml_constants.py
@@ -1,1 +1,0 @@
-CHECKER_ID = "xml_xosc"
