+++ tests/test_utils.py
@@ -1,0 +1,46 @@
+import os
+import sys
+import pytest
+from typing import List
+import main
+from qc_openscenario import constants, checks
+from qc_baselib import Configuration, Result
+
+
+CONFIG_FILE_PATH = "bundle_config.xml"
+REPORT_FILE_PATH = "xosc_bundle_report.xqar"
+
+
+def create_test_config(target_file_path: str):
+    test_config = Configuration()
+    test_config.set_config_param(name="XoscFile", value=target_file_path)
+    test_config.register_checker_bundle(checker_bundle_name=constants.BUNDLE_NAME)
+    test_config.set_checker_bundle_param(
+        checker_bundle_name=constants.BUNDLE_NAME,
+        name="resultFile",
+        value=REPORT_FILE_PATH,
+    )
+
+    test_config.write_to_file(CONFIG_FILE_PATH)
+
+
+def launch_main(monkeypatch):
+    monkeypatch.setattr(
+        sys,
+        "argv",
+        [
+            "main.py",
+            "-c",
+            CONFIG_FILE_PATH,
+        ],
+    )
+    main.main()
+
+
+def cleanup_files():
+    os.remove(REPORT_FILE_PATH)
+    os.remove(CONFIG_FILE_PATH)
+
+
+def get_issues_by_rule_name(checker_result: Result, rule_name: str) -> List:
+    return [x for x in checker_result.issues if x.rule_uid.split(":")[-1] == rule_name]
